Create separate component for handling success messages
