+++ src/components/ErrorHandler.js
@@ -1,5 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
+// STORE
 import { clearLoadArticlesError, clearAddArticleError, clearEditArticleError, clearRemoveArticleError } from 'redux/modules/articlesModule';
 import { clearLoadConversationError } from 'redux/modules/conversationModule';
 // COMPONENTS
