+++ src/components/index.js
@@ -5,4 +5,5 @@ export ErrorHandler from './ErrorHandler';
 export ErrorSnackbar from './ErrorSnackbar';
 export LoadingScreen from './LoadingScreen';
 export Spinner from './Spinner';
+export SuccessHandler from './SuccessHandler';
 export SuccessSnackbar from './SuccessSnackbar';
